Now shadow dir is kept when running with keep-incomplete

+++ src/snakemake/dag.py
@@ -784,7 +784,7 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
                         rule=job.rule,
                     )
 
-    def unshadow_output(self, job, only_log=False):
+    def unshadow_output(self, job, only_log=False, keep_shadow_dir=False):
         """Move files from shadow directory to real output paths."""
         if not job.shadow_dir or not job.output:
             return
@@ -804,7 +804,14 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
             if os.path.realpath(shadow_output) == os.path.realpath(real_output):
                 continue
             shutil.move(shadow_output, real_output)
-        shutil.rmtree(job.shadow_dir)
+        if (keep_shadow_dir and os.listdir(job.shadow_dir)):
+            print(
+                f"Keeping shadow directory. " +
+                f"at path: {job.shadow_dir}.\n" + 
+                "Run snakemake with --cleanup-shadow to clean up shadow directories."
+            )
+        else:
+            shutil.rmtree(job.shadow_dir)
 
     def check_periodic_wildcards(self, job):
         """Raise an exception if a wildcard of the given job appears to be periodic,
